+++ src/main/java/com/example/samsung_delivery/repository/CouponHistoryRepository.java
@@ -2,7 +2,11 @@ package com.example.samsung_delivery.repository;
 
 
 import com.example.samsung_delivery.entity.CouponHistory;
+import com.example.samsung_delivery.enums.CouponStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -16,4 +20,9 @@ public interface CouponHistoryRepository extends JpaRepository<CouponHistory , L
     long countByCoupon_Id(Long couponId);
 
     Optional<CouponHistory> findByUser_IdAndCoupon_Id(Long userId, Long couponId);
+
+    @Modifying
+    @Query(value = "update CouponHistory ch set ch.status = :status where ch.coupon.id in " +
+            "(select c.id from Coupon c where c.expiredAt < now())")
+    void updateStatus(@Param("status")CouponStatus status);
 }
